Update readme.md
I added more info about the parameter when you deploy AVS monitoring. Upon testing this to customer environment, we had difficulty in finding the Resource ID of AVS Private Cloud. I added in this guide the format for the resource ID. I don't have permission to upload an image, I was also planning to put some screenshot as an example.